tweak params

+++ Client.hpp
@@ -12,7 +12,7 @@ class Client {
 public:
     Client();
     ~Client();
-    void registration(double revealed_per_interval, int tags_each_reveal, int period_length);
+    void registration(int revealed_per_interval, int tags_each_reveal, int period_length);
     void reveal(float percentage);
     void payment();
     int cc_bytes; //communication cost in bytes
